ref: Refactor loader settings into separate file
