+++ static/app/views/settings/project/projectKeys/details/keySettings.tsx
@@ -1,4 +1,4 @@
-import {Fragment, useCallback, useEffect, useState} from 'react';
+import {Fragment, useCallback} from 'react';
 
 import {
   addErrorMessage,
@@ -11,24 +11,14 @@ import Confirm from 'sentry/components/confirm';
 import DateTime from 'sentry/components/dateTime';
 import FieldGroup from 'sentry/components/forms/fieldGroup';
 import BooleanField from 'sentry/components/forms/fields/booleanField';
-import SelectField from 'sentry/components/forms/fields/selectField';
 import TextField from 'sentry/components/forms/fields/textField';
 import Form from 'sentry/components/forms/form';
-import ExternalLink from 'sentry/components/links/externalLink';
-import {
-  Panel,
-  PanelAlert,
-  PanelBody,
-  PanelFooter,
-  PanelHeader,
-} from 'sentry/components/panels';
-import TextCopyInput from 'sentry/components/textCopyInput';
-import {t, tct} from 'sentry/locale';
+import {Panel, PanelAlert, PanelBody, PanelHeader} from 'sentry/components/panels';
+import {t} from 'sentry/locale';
 import {Organization} from 'sentry/types';
-import getDynamicText from 'sentry/utils/getDynamicText';
-import handleXhrErrorResponse from 'sentry/utils/handleXhrErrorResponse';
 import useApi from 'sentry/utils/useApi';
 import KeyRateLimitsForm from 'sentry/views/settings/project/projectKeys/details/keyRateLimitsForm';
+import {LoaderSettings} from 'sentry/views/settings/project/projectKeys/details/loaderSettings';
 import ProjectKeyCredentials from 'sentry/views/settings/project/projectKeys/projectKeyCredentials';
 import {ProjectKey} from 'sentry/views/settings/project/projectKeys/types';
 
@@ -42,48 +32,11 @@ type Props = {
   };
 };
 
-export enum DynamicSDKLoaderOption {
-  HAS_DEBUG = 'hasDebug',
-  HAS_PERFORMANCE = 'hasPerformance',
-  HAS_REPLAY = 'hasReplay',
-}
-
-export const sdkLoaderOptions = {
-  [DynamicSDKLoaderOption.HAS_PERFORMANCE]: {
-    label: t('Enable Performance Monitoring'),
-    requiresV7: true,
-  },
-  [DynamicSDKLoaderOption.HAS_REPLAY]: {
-    label: t('Enable Session Replay'),
-    requiresV7: true,
-  },
-  [DynamicSDKLoaderOption.HAS_DEBUG]: {
-    label: t('Enable Debug Bundles'),
-    requiresV7: false,
-  },
-};
-
 export function KeySettings({onRemove, organization, params, data}: Props) {
   const api = useApi();
-  const [browserSdkVersion, setBrowserSdkVersion] = useState(data.browserSdkVersion);
-  const [dynamicSDKLoaderOptions, setDynamicSDKLoaderOptions] = useState(
-    data.dynamicSdkLoaderOptions
-  );
-
-  useEffect(() => {
-    setBrowserSdkVersion(data.browserSdkVersion);
-  }, [data.browserSdkVersion]);
-
-  useEffect(() => {
-    setDynamicSDKLoaderOptions(data.dynamicSdkLoaderOptions);
-  }, [data.dynamicSdkLoaderOptions]);
 
   const {keyId, projectId} = params;
   const apiEndpoint = `/projects/${organization.slug}/${projectId}/keys/${keyId}/`;
-  const loaderLink = getDynamicText({
-    value: data.dsn.cdn,
-    fixed: '__JS_SDK_LOADER_URL__',
-  });
 
   const handleRemove = useCallback(async () => {
     addLoadingMessage(t('Revoking key\u2026'));
@@ -102,96 +55,6 @@ export function KeySettings({onRemove, organization, params, data}: Props) {
       addErrorMessage(t('Unable to revoke key'));
     }
   }, [organization, api, onRemove, keyId, projectId]);
-
-  const handleToggleDynamicSDKLoaderOption = useCallback(
-    async <T extends typeof dynamicSDKLoaderOptions, K extends keyof T>(
-      dynamicSdkLoaderOption: K,
-      value: T[K]
-    ) => {
-      const newDynamicSdkLoaderOptions = Object.keys(dynamicSDKLoaderOptions).reduce(
-        (acc, key) => {
-          if (key === dynamicSdkLoaderOption) {
-            return {...acc, [key]: value};
-          }
-          return {...acc, [key]: dynamicSDKLoaderOptions[key]};
-        },
-        {}
-      );
-
-      addLoadingMessage();
-
-      try {
-        const response = await api.requestPromise(apiEndpoint, {
-          method: 'PUT',
-          data: {
-            dynamicSdkLoaderOptions: newDynamicSdkLoaderOptions,
-          },
-        });
-
-        setDynamicSDKLoaderOptions(response.dynamicSdkLoaderOptions);
-
-        addSuccessMessage(t('Successfully updated dynamic SDK loader configuration'));
-      } catch (error) {
-        const message = t('Unable to updated dynamic SDK loader configuration');
-        handleXhrErrorResponse(message)(error);
-        addErrorMessage(message);
-      }
-    },
-    [api, apiEndpoint, dynamicSDKLoaderOptions, setDynamicSDKLoaderOptions]
-  );
-
-  const handleUpdateBrowserSDKVersion = useCallback(
-    async (newBrowserSDKVersion: typeof browserSdkVersion) => {
-      addLoadingMessage();
-
-      const apiData: {
-        browserSdkVersion: typeof browserSdkVersion;
-        dynamicSdkLoaderOptions?: Partial<Record<DynamicSDKLoaderOption, boolean>>;
-      } = {
-        browserSdkVersion: newBrowserSDKVersion,
-      };
-
-      const shouldRestrictDynamicSdkLoaderOptions =
-        !sdkVersionSupportsPerformanceAndReplay(newBrowserSDKVersion);
-
-      if (shouldRestrictDynamicSdkLoaderOptions) {
-        // Performance & Replay are not supported before 7.x
-        const newDynamicSdkLoaderOptions = {
-          ...dynamicSDKLoaderOptions,
-          hasPerformance: false,
-          hasReplay: false,
-        };
-
-        apiData.dynamicSdkLoaderOptions = newDynamicSdkLoaderOptions;
-      }
-
-      try {
-        const response = await api.requestPromise(apiEndpoint, {
-          method: 'PUT',
-          data: apiData,
-        });
-
-        setBrowserSdkVersion(response.browserSdkVersion);
-
-        if (shouldRestrictDynamicSdkLoaderOptions) {
-          setDynamicSDKLoaderOptions(response.dynamicSdkLoaderOptions);
-        }
-
-        addSuccessMessage(t('Successfully updated SDK version'));
-      } catch (error) {
-        const message = t('Unable to updated SDK version');
-        handleXhrErrorResponse(message)(error);
-        addErrorMessage(message);
-      }
-    },
-    [
-      api,
-      apiEndpoint,
-      setBrowserSdkVersion,
-      setDynamicSDKLoaderOptions,
-      dynamicSDKLoaderOptions,
-    ]
-  );
 
   return (
     <Access access={['project:write']}>
@@ -241,97 +104,13 @@ export function KeySettings({onRemove, organization, params, data}: Props) {
           <Panel>
             <PanelHeader>{t('JavaScript Loader')}</PanelHeader>
             <PanelBody>
-              <FieldGroup
-                help={tct(
-                  'Copy this script into your website to setup your JavaScript SDK without any additional configuration. [link]',
-                  {
-                    link: (
-                      <ExternalLink href="https://docs.sentry.io/platforms/javascript/install/lazy-load-sentry/">
-                        What does the script provide?
-                      </ExternalLink>
-                    ),
-                  }
-                )}
-                inline={false}
-                flexibleControlStateSize
-              >
-                <TextCopyInput>
-                  {`<script src='${loaderLink}' crossorigin="anonymous"></script>`}
-                </TextCopyInput>
-              </FieldGroup>
-              <SelectField
-                name="browserSdkVersion"
-                options={
-                  data.browserSdk
-                    ? data.browserSdk.choices.map(([value, label]) => ({
-                        value,
-                        label,
-                      }))
-                    : []
-                }
-                value={browserSdkVersion}
-                onChange={value => handleUpdateBrowserSDKVersion(value)}
-                placeholder={t('4.x')}
-                allowClear={false}
-                disabled={!hasAccess}
-                help={t(
-                  'Select the version of the SDK that should be loaded. Note that it can take a few minutes until this change is live.'
-                )}
+              <LoaderSettings
+                organizationId={organization.slug}
+                keyId={params.keyId}
+                projectId={params.projectId}
+                projectKey={data}
               />
             </PanelBody>
-
-            <PanelFooter>
-              {Object.entries(sdkLoaderOptions).map(([key, value]) => {
-                const sdkLoaderOption = Object.keys(dynamicSDKLoaderOptions).find(
-                  dynamicSdkLoaderOption => dynamicSdkLoaderOption === key
-                );
-
-                if (!sdkLoaderOption) {
-                  return null;
-                }
-
-                return (
-                  <BooleanField
-                    label={value.label}
-                    key={key}
-                    name={key}
-                    value={
-                      value.requiresV7 &&
-                      !sdkVersionSupportsPerformanceAndReplay(browserSdkVersion)
-                        ? false
-                        : dynamicSDKLoaderOptions[sdkLoaderOption]
-                    }
-                    onChange={() =>
-                      handleToggleDynamicSDKLoaderOption(
-                        sdkLoaderOption as DynamicSDKLoaderOption,
-                        !dynamicSDKLoaderOptions[sdkLoaderOption]
-                      )
-                    }
-                    disabled={
-                      !hasAccess ||
-                      (value.requiresV7 &&
-                        !sdkVersionSupportsPerformanceAndReplay(browserSdkVersion))
-                    }
-                    help={
-                      value.requiresV7 &&
-                      !sdkVersionSupportsPerformanceAndReplay(browserSdkVersion)
-                        ? t('Only available in SDK version 7.x and above')
-                        : key === DynamicSDKLoaderOption.HAS_REPLAY &&
-                          dynamicSDKLoaderOptions[sdkLoaderOption]
-                        ? t(
-                            'When using Replay, the loader will load the ES6 bundle instead of the ES5 bundle.'
-                          )
-                        : undefined
-                    }
-                    disabledReason={
-                      !hasAccess
-                        ? t('You do not have permission to edit this setting')
-                        : undefined
-                    }
-                  />
-                );
-              })}
-            </PanelFooter>
           </Panel>
 
           <Panel>
@@ -383,8 +162,4 @@ export function KeySettings({onRemove, organization, params, data}: Props) {
       )}
     </Access>
   );
-}
-
-function sdkVersionSupportsPerformanceAndReplay(sdkVersion: string): boolean {
-  return sdkVersion === 'latest' || sdkVersion === '7.x';
 }
